Test: BoardService, BoardCommentService Test Code add
Fix: deleteBoard method error correction

+++ src/test/java/com/wolves/mainproject/service/BoardServiceTest.java
@@ -6,13 +6,16 @@ import com.wolves.mainproject.domain.board.comment.BoardComment;
 import com.wolves.mainproject.domain.board.comment.BoardCommentRepository;
 import com.wolves.mainproject.domain.board.content.BoardContent;
 import com.wolves.mainproject.domain.board.content.BoardContentRepository;
+import com.wolves.mainproject.domain.board.like.BoardLike;
+import com.wolves.mainproject.domain.board.like.BoardLikeRepository;
 import com.wolves.mainproject.domain.user.User;
 import com.wolves.mainproject.domain.user.UserRepository;
 import com.wolves.mainproject.dto.request.board.BoardRequestDto;
+import com.wolves.mainproject.dto.response.BoardResponseDto;
 import com.wolves.mainproject.dto.response.GetBoardResponseDto;
+import com.wolves.mainproject.exception.board.BoardPageNotFoundException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -22,18 +25,12 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
-/**
- * 단위 테스트 (Service와 관련된 애들만 메모리에 띄우면 됨.)
- * BoardRepository => 가짜 객체로 만들 수 있음.
- *
- */
 
 
 //@ExtendWith(MockitoExtension.class)
 @SpringBootTest
 public class BoardServiceTest {
 
-    @Mock
     @Autowired
     private BoardRepository boardRepository;
 
@@ -46,6 +43,9 @@ public class BoardServiceTest {
     @Autowired
     private BoardCommentRepository boardCommentRepository;
 
+    @Autowired
+    private BoardLikeRepository boardLikeRepository;
+
 
     @BeforeEach
     void createBoard() {
@@ -54,29 +54,66 @@ public class BoardServiceTest {
         boardRepository.save(board);
         BoardContent boardContent = BoardContent.builder().board(board).content("boardContent").build();
         boardContentRepository.save(boardContent);
+        BoardComment boardComment = BoardComment.builder().board(board).status(true).user(user).refer(null).content("testComment").build();
+        boardCommentRepository.save(boardComment);
+        BoardComment boardReComment = BoardComment.builder().board(board).status(true).user(user).refer(boardComment).content("testReComment").build();
+        boardCommentRepository.save(boardReComment);
     }
 
-//    @BeforeEach
-//    void createComment(){
-//        User user = userRepository.findById(1L).orElseThrow();
-//        Board board = boardRepository.findById(1L).orElseThrow();
-//        BoardComment boardComment = BoardComment.builder().board(board).content("test입니다").user(user).build();
-//        boardCommentRepository.save(boardComment);
-//
-//    }
+    @BeforeEach
+    void createBoard2() {
+        User user = userRepository.findById(1L).orElseThrow();
+        Board board = Board.builder().title("board2").user(user).build();
+        boardRepository.save(board);
+        BoardContent boardContent = BoardContent.builder().board(board).content("boardContent2").build();
+        boardContentRepository.save(boardContent);
+
+    }
+
+
+    @Test
+    void getAllBoard(){
+        //when
+        List<Board> boards = boardRepository.findAll();
+        //then
+        assertEquals(2,boards.size());
+    }
+
+    @Test
+    void searchBoard() {
+        //given
+        String search = "2";
+        //when
+        List<Board> boardsList = boardRepository.findByTitleContainingOrUserNicknameContaining(search);
+        assert boardsList != null;
+        //then
+        assertEquals(1, boardsList.size());
+    }
+
+    @Test
+    void getBoardDetails() {
+        //given
+        long boardId = 1L;
+        //when
+        Board board = boardRepository.findById(boardId).orElseThrow(BoardPageNotFoundException::new);
+        BoardContent boardContent = boardContentRepository.findById(boardId).orElseThrow(BoardPageNotFoundException::new);
+        List<BoardComment> boardCommentList = boardCommentRepository.findByBoard(board);
+        List<GetBoardResponseDto> getBoardResponseDtos = boardCommentList.stream().map(boardComment -> new GetBoardResponseDto(boardComment)).toList();
+        new BoardResponseDto(board,boardContent,getBoardResponseDtos);
+        //then
+        assertEquals(board.getUser().getId(),boardContent.getBoard().getUser().getId());
+        assertEquals(getBoardResponseDtos.get(0).getId(),getBoardResponseDtos.get(1).getReferComment());
+    }
 
 
     @Test
     void updateBoard() {
         //Given
         User user = userRepository.findById(1L).orElseThrow();
-        Board board = boardRepository.findById(1L).orElseThrow();
+        Board board = boardRepository.findById(1L).orElseThrow(BoardPageNotFoundException::new);
         if (user.equals(board.getUser())) {
             BoardRequestDto dto = BoardRequestDto.builder().title("updateboard").content("updatecontent").build();
             BoardContent boardContent = boardContentRepository.findByBoard(board);
-
-        List<BoardComment> boardCommentList = boardCommentRepository.findByBoard(board);
-        List<GetBoardResponseDto> getBoardResponseDtos = boardCommentList.stream().map(boardComment -> new GetBoardResponseDto(boardComment)).toList();
             //when
             board.update(dto);
             Board boardPS = boardRepository.save(board);
@@ -84,8 +121,8 @@ public class BoardServiceTest {
             BoardContent boardContentPS = boardContentRepository.save(boardContent);
 
             //then
-            assertEquals("updateboard", dto.getTitle());
-            assertEquals("updatecontent", dto.getContent());
+            assertEquals("updateboard", boardPS.getTitle());
+            assertEquals("updatecontent", boardContentPS.getContent());
         }
     }
 
@@ -93,13 +130,25 @@ public class BoardServiceTest {
     void deleteBoard() {
         //given
         long deleteId = 1L;
-        Board board = boardRepository.findById(deleteId).orElseThrow();
+        Board board = boardRepository.findById(deleteId).orElseThrow(BoardPageNotFoundException::new);
         //when
-        boardRepository.delete(board);
-
+        if (deleteId==board.getUser().getId())
+            boardRepository.delete(board);
         //then
         assertNull(boardRepository.findById(deleteId).orElse(null));
     }
-
+    @Test
+    void findLikeBoardList(){
+        //given
+        User user = userRepository.findById(1L).orElseThrow();
+        Board board = boardRepository.findById(1L).orElseThrow();
+        boardLikeRepository.save(BoardLike.builder().board(board).user(user).build());
+        Board board2 = boardRepository.findById(2L).orElseThrow();
+        boardLikeRepository.save(BoardLike.builder().board(board2).user(user).build());
+        //when
+        List<BoardLike> boardlikes = boardLikeRepository.findByUser(user);
+        assert boardlikes != null;
+        //then
+        assertEquals(2,boardlikes.size());
+    }
 }
-
